Marge the bit field and alphabitfile looks like use less to separate in another class. tried to use or in switch expression fix handling more than just 10 bit. might able to opthalmic it farther due to the simplification of the bit shifting logic

+++ Src/ImageQT/Decoder/BMP/Models/ColorReader/BitFieldColorReader.cs
@@ -42,10 +42,11 @@ internal class BitFieldColorReader : BaseColorReader
         // 32
         if (HeaderDetails.BitDepth == 32)
         {
+            // 32 bit has some issue
             var value = BinaryPrimitives.ReadInt32LittleEndian(pixel);
-            var r = MapTo8Bits((byte)((value & HeaderDetails.RedMask) >> _redShift), _redMaskSize);
-            var g = MapTo8Bits((byte)((value & HeaderDetails.GreenMask) >> _greenShift), _greenMaskSize);
-            var b = MapTo8Bits((byte)((value & HeaderDetails.BlueMask) >> _blueShift), _blueMaskSize);
+            var r = MapTo8Bits(((value & HeaderDetails.RedMask) >> _redShift), _redMaskSize);
+            var g = MapTo8Bits(((value & HeaderDetails.GreenMask) >> _greenShift), _greenMaskSize);
+            var b = MapTo8Bits(((value & HeaderDetails.BlueMask) >> _blueShift), _blueMaskSize);
             result[writingIndex++] = new Pixels(r, g, b);
         }
     }
